Move error conversion to peripheral.rs
